+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixed
 
 - Gracefully handle exceptions thrown from block render code
+- Ensure shutdown on world leave
 
 ## [1.0.2] - 2025-09-26